+++ src/sqlalchemy_declarative_extensions/function/base.py
@@ -24,6 +24,19 @@ class Function:
     language: str = "sql"
     schema: str | None = None
 
+    @classmethod
+    def from_unknown_function(cls, f: Function) -> Self:
+        if isinstance(f, cls):
+            return f
+
+        return cls(
+            name=f.name,
+            definition=f.definition,
+            language=f.language,
+            schema=f.schema,
+            returns=f.returns,
+        )
+
     @property
     def qualified_name(self):
         return qualify_name(self.schema, self.name)
@@ -31,14 +44,17 @@ class Function:
     def normalize(self) -> Self:
         raise NotImplementedError()  # pragma: no cover
 
-    def to_sql_create(self, replace=False):
+    def to_sql_create(self) -> list[str]:
         raise NotImplementedError()
 
-    def to_sql_update(self):
-        return self.to_sql_create(replace=True)
+    def to_sql_update(self) -> list[str]:
+        return [
+            *self.to_sql_drop(),
+            *self.to_sql_create(),
+        ]
 
-    def to_sql_drop(self):
-        return f"DROP FUNCTION {self.qualified_name}();"
+    def to_sql_drop(self) -> list[str]:
+        return [f"DROP FUNCTION {self.qualified_name}();"]
 
     def with_name(self, name: str):
         return replace(self, name=name)